strings test passes
