+++ crates/test/src/cpp.rs
@@ -96,6 +96,7 @@ impl LanguageMethods for Cpp17 {
         .arg(&compile.bindings_dir)
         .arg("-I")
         .arg(helper_dir.to_str().unwrap().to_string())
+        .arg("-fno-exceptions")
         .arg("-Wall")
         .arg("-Wextra")
         .arg("-Werror")
@@ -116,6 +117,9 @@ impl LanguageMethods for Cpp17 {
             )))
             .arg("-I")
             .arg(&compile.bindings_dir)
+            .arg("-I")
+            .arg(helper_dir.to_str().unwrap().to_string())
+            .arg("-fno-exceptions")
             .arg("-Wall")
             .arg("-Wextra")
             .arg("-Werror")
