+++ tests/runtime/strings/runner.cpp
@@ -6,22 +6,22 @@
 #include <stdio.h>
 #include <runner_cpp.h>
 
-void assert_str(std::string_view str, const char* expected) {
+void assert_str(wit::string const& str, const char* expected) {
   size_t expected_len = strlen(expected);
   assert(str.size() == expected_len);
   assert(memcmp(str.data(), expected, expected_len) == 0);
 }
 
 int main() {
-    test::strings::to_test::take_basic("latin utf16");
+    test::strings::to_test::TakeBasic("latin utf16");
 
-    let str2 = test::strings::to_test::return_unicode();
+    auto str2 = test::strings::to_test::ReturnUnicode();
     assert_str(str2, "🚀🚀🚀 𠈄𓀀");
 
-    let str3 = test::strings::to_test::return_empty();
+    auto str3 = test::strings::to_test::ReturnEmpty();
     assert_str(str3, "");
 
-    let str5 = test::strings::to_test::roundtrip("🚀🚀🚀 𠈄𓀀");
+    auto str5 = test::strings::to_test::Roundtrip("🚀🚀🚀 𠈄𓀀");
     assert_str(str5, "🚀🚀🚀 𠈄𓀀");
     
     return 0;
